Major Changes
Added percentages to User Data
Need to add a transparent image over the background
Need to go and change all button interactions to be on release.
Need to change UI of the User Data. Change it to card style
Need to put a margin on the User activity page.
Need to get a more meaningful UI to the User Activity page.